move HDF5 saving and loading into an extension
